UrbanClap/backend/requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
Celery==5.3.6
redis==5.0.1
Werkzeug==3.0.1
PyJWT==2.8.0
SQLAlchemy==2.0.27
python-dateutil==2.8.2
